correct some small issues

+++ Models/Measurement.cs
@@ -1,10 +1,10 @@
 ﻿namespace ScyllaDBDemo.Models;
 
-public class Measurement(Sensor sensor, double value)
+public class Measurement(Sensor sensor, float value)
 {
     public Guid SensorId { get; set;} = sensor.PetId;
 
     public DateTimeOffset Timestamp { get; set; } = new(DateTime.UtcNow);
 
-    public double Value { get; set; } = value;
+    public float Value { get; set; } = value;
 }
